10_scraping/01_structure_data_scraper/requirements.txt
crawl4ai>=0.4
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
